emotion-detection-server/emotion_api/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
Pillow==10.2.0
numpy==1.24.3
opencv-python==4.8.0.76
tensorflow-cpu==2.12.0
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
matplotlib==3.7.2 